+++ burnxp/src/train.rs
@@ -1,5 +1,6 @@
 use burn::{
     data::dataloader::DataLoaderBuilder,
+    lr_scheduler::linear::LinearLrSchedulerConfig,
     optim::AdamConfig,
     prelude::*,
     record::{CompactRecorder, Recorder},
@@ -114,7 +115,7 @@ pub fn train<B: AutodiffBackend>(artifact_dir: PathBuf, config: TrainingConfig, 
                 model
             },
             config.optimizer.init(),
-            config.learning_rate,
+            LinearLrSchedulerConfig::new(config.learning_rate, config.learning_rate/10., config.num_epochs).init(),
         );
 
     let model_trained = learner.fit(dataloader_train, dataloader_valid);